+++ changelog.md
@@ -3,6 +3,7 @@
 ## 0.6.5
 - Clean up linting errors
 - Update dependencies
+- Add debug logs for serializing and deserializing structures
 
 ## 0.6.4
 - Add bool as a primitive type